requirements.txt
google-api-python-client==2.151.0
pandas==2.2.3
XlsxWriter==3.2.0
asyncio==3.4.3
playwright==1.49.0
glom==24.11.0
python-dotenv==1.0.1
isodate==0.7.2
tqdm==4.67.1